# AI Customer Service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/orders/": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create Order",
        "description": "创建订单（当前登录用户）",
        "operationId": "create_order_api_orders__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_OrderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List Orders",
        "description": "分页查询当前用户订单列表",
        "operationId": "list_orders_api_orders__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "paid",
                    "shipped",
                    "completed",
                    "cancelled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态筛选",
              "title": "Status"
            },
            "description": "按状态筛选"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "分页偏移量",
              "default": 0,
              "title": "Skip"
            },
            "description": "分页偏移量"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_OrderResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/order-no/{order_no}": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get Order",
        "description": "根据订单号查询订单详情（当前登录用户）",
        "operationId": "get_order_api_orders_order_no__order_no__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_OrderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "orders"
        ],
        "summary": "Update Order",
        "description": "根据订单号更新订单（状态/收货信息/备注）",
        "operationId": "update_order_api_orders_order_no__order_no__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order No"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_OrderResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "orders"
        ],
        "summary": "Delete Order",
        "description": "根据订单号删除订单（物理删除，生产可改为逻辑删除）",
        "operationId": "delete_order_api_orders_order_no__order_no__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Create Ticket",
        "description": "创建工单（当前登录用户的反馈 / 投诉）",
        "operationId": "create_ticket_api_tickets__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TicketResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "List Tickets",
        "description": "分页查询当前用户的工单列表",
        "operationId": "list_tickets_api_tickets__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "open",
                    "processing",
                    "closed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态筛选",
              "title": "Status"
            },
            "description": "按状态筛选"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "分页偏移量",
              "default": 0,
              "title": "Skip"
            },
            "description": "分页偏移量"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_TicketResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Get Ticket",
        "description": "获取当前用户的某个工单详情",
        "operationId": "get_ticket_api_tickets__ticket_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TicketResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "tickets"
        ],
        "summary": "Update Ticket",
        "description": "更新当前用户工单（状态 / 分配人 / 处理结果）",
        "operationId": "update_ticket_api_tickets__ticket_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TicketResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tickets"
        ],
        "summary": "Delete Ticket",
        "description": "删除当前用户工单（物理删除，生产可改为仅标记关闭）",
        "operationId": "delete_ticket_api_tickets__ticket_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_List_OrderResponse__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OrderResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[OrderResponse]]"
      },
...
      "ApiResponse_List_TicketResponse__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TicketResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[TicketResponse]]"
      },
      "ApiResponse_OrderResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[OrderResponse]"
      },
...
      "ApiResponse_TicketResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ApiResponse[TicketResponse]"
      },
...
      "ApiResponse_bool_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Success"
          },
          "data": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "title": "ApiResponse[bool]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "下单用户ID"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItem-Input"
            },
            "type": "array",
            "title": "Items",
            "description": "商品列表"
          },
          "total_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Total Amount",
            "description": "订单总金额"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "shipped",
              "completed",
              "cancelled"
            ],
            "title": "Status",
            "description": "订单状态",
            "default": "pending"
          },
          "receiver_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Receiver Name",
            "description": "收货人姓名"
          },
          "receiver_phone": {
            "type": "string",
            "maxLength": 20,
            "title": "Receiver Phone",
            "description": "收货人电话"
          },
          "shipping_address": {
            "type": "string",
            "title": "Shipping Address",
            "description": "收货地址"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "订单备注"
          },
          "order_no": {
            "type": "string",
            "maxLength": 50,
            "title": "Order No",
            "description": "订单号，唯一"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "items",
          "total_amount",
          "receiver_name",
          "receiver_phone",
          "shipping_address",
          "order_no"
        ],
        "title": "OrderCreate",
        "description": "创建订单请求"
      },
...
      "OrderUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "pending",
                  "paid",
                  "shipped",
                  "completed",
                  "cancelled"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "新的订单状态"
          },
          "receiver_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Receiver Name",
            "description": "新的收货人姓名"
          },
          "receiver_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Receiver Phone",
            "description": "新的收货人电话"
          },
          "shipping_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Address",
            "description": "新的收货地址"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "新的订单备注"
          }
        },
        "type": "object",
        "title": "OrderUpdate",
        "description": "更新订单请求"
      },
...
      "TicketCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "提交用户ID"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id",
            "description": "关联的订单ID"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "关联的会话ID"
          },
          "type": {
            "type": "string",
            "enum": [
              "complaint",
              "return",
              "exchange",
              "other"
            ],
            "title": "Type",
            "description": "工单类型"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "工单标题"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "详细描述"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "processing",
              "closed"
            ],
            "title": "Status",
            "description": "工单状态",
            "default": "open"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To",
            "description": "分配给的处理人"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution",
            "description": "处理结果"
          },
          "ticket_no": {
            "type": "string",
            "maxLength": 50,
            "title": "Ticket No",
            "description": "工单号，唯一"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "type",
          "description",
          "ticket_no"
        ],
        "title": "TicketCreate",
        "description": "创建工单请求"
      },
...
      "TicketUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "open",
                  "processing",
                  "closed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "新的工单状态"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To",
            "description": "新的处理人"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution",
            "description": "处理结果"
          }
        },
        "type": "object",
        "title": "TicketUpdate",
        "description": "更新工单请求"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/users/token"
          }
        }
      }
    }
  }
}